exp1/requirement.txt
需要的库

torch

transformer

sklearn

time

json

numpy